% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellTYpeNameExtend.R, R/GetCellTypeName.R
\name{as_matrix}
\alias{as_matrix}
\title{Change to matrix}
\usage{
as_matrix(mat)

as_matrix(mat)
}
\arguments{
\item{mat}{data input}
}
\value{
A numeric vector of log density.

A numeric vector of log density.
}
\description{
Creating poisson vector.

Creating poisson vector.
}
\details{
Input an integer and return the log density of a poisson distribution with lambda equals the input integer

Input an integer and return the log density of a poisson distribution with lambda equals the input integer
}
