% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellTYpeNameExtend.R
\name{GetCellTYpeNameExtend}
\alias{GetCellTYpeNameExtend}
\title{Get cell type name which can not find by ROC method}
\usage{
GetCellTYpeNameExtend(anno, new_sample_data, new_meta)
}
\arguments{
\item{anno}{Data anno.}

\item{new_sample_data}{Pse data.}

\item{new_meta}{Pse meta.}
}
\value{
A dataframe contains annotation of cell type name
}
\description{
Get cell type name which can not find by ROC method.
}
