% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterCluster.R
\name{IterCluster}
\alias{IterCluster}
\title{Iteractive clustering}
\usage{
IterCluster(
  norm.dat,
  select.cells = colnames(norm.dat),
  prefix = NULL,
  k.nn,
  max.dim,
  de.param = de_param(),
  merge.type = c("undirectional", "directional"),
  split.size = 10,
  result = NULL,
  method = "auto"
)
}
\arguments{
\item{norm.dat}{data input}

\item{select.cells}{data ref}

\item{prefix}{meta input}

\item{k.nn}{meta input}

\item{max.dim}{meta input}

\item{de.param}{The differential gene expression threshold.}

\item{merge.type}{Determine if the DE gene score threshold should be applied to combined de.score, or de.score for up and down directions separately.}

\item{split.size}{meta input}

\item{result}{meta input}

\item{method}{meta input}
}
\value{
Cluster result.
}
\description{
Iterative clustering algorithm for single cell RNAseq dataset.
}
