% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OneStepCluster.R
\name{OneStepCluster}
\alias{OneStepCluster}
\title{One step cluster}
\usage{
OneStepCluster(
  norm.dat,
  select.cells = colnames(norm.dat),
  counts = NULL,
  method = c("louvain", "leiden", "ward.D", "kmeans"),
  vg.padj.th = 0.5,
  dim.method = c("pca", "WGCNA"),
  max.dim = 20,
  rm.eigen = NULL,
  rm.th = 0.7,
  de.param = de_param(),
  merge.type = c("undirectional", "directional"),
  maxGenes = 3000,
  sampleSize = 40000,
  max.cl.size = 500,
  k.nn = 50,
  prefix = NULL,
  verbose = FALSE,
  regress.x = NULL
)
}
\arguments{
\item{norm.dat}{data input}

\item{select.cells}{data ref}

\item{counts}{meta input}

\item{method}{meta input}

\item{vg.padj.th}{meta input}

\item{dim.method}{meta input}

\item{max.dim}{meta input}

\item{rm.eigen}{meta input}

\item{rm.th}{meta input}

\item{de.param}{meta input}

\item{merge.type}{meta input}

\item{maxGenes}{meta input}

\item{sampleSize}{meta input}

\item{max.cl.size}{meta input}

\item{k.nn}{meta input}

\item{prefix}{meta input}

\item{verbose}{meta input}

\item{regress.x}{meta input}
}
\value{
A numeric vector of log density.
}
\description{
Creating poisson vector.
}
\details{
Input an integer and return the log density of a poisson distribution with lambda equals the input integer
}
