% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunConsensus1.R
\name{RunConsensus1}
\alias{RunConsensus1}
\title{Run consensus1}
\usage{
RunConsensus1(
  norm.dat,
  select.cells = colnames(norm.dat),
  k.nn = 15,
  max.dim = 20,
  output_dir = "subsample_result",
  mc.cores = 1,
  de.param = de_param(),
  merge.type = c("undirectional", "directional"),
  override = FALSE,
  init.result = NULL,
  cut.method = "auto",
  ...
)
}
\arguments{
\item{norm.dat}{Single cell expression data input.}

\item{select.cells}{Numeric, use all or part of the cells for clustering, cells can be sampled to save running time in multi-round clustering. Using all cells by default.}

\item{k.nn}{K-nearst neighbors used for clustering. Default 15.}

\item{max.dim}{The number of PCA dimensions retained. Default 20.}

\item{output_dir}{File directory to store clustering results.}

\item{mc.cores}{Number of cores used for parallel processing.}

\item{de.param}{The differential gene expression threshold.}

\item{merge.type}{Determine if the DE gene score threshold should be applied to combined de.score, or de.score for up and down directions separately.}

\item{override}{Clustering method. It can be "auto", "louvain", "hclust".}

\item{init.result}{Clustering method. It can be "auto", "louvain", "hclust".}

\item{cut.method}{Clustering method. It can be "auto", "louvain", "hclust".}

\item{...}{Other parameters.}
}
\value{
A list with two elements: cl: cluster membership for each cell markers: top markers that seperate clusters
}
\description{
Using consensus1 method to clust single cell data
}
