% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFindDEGene.R
\name{RunFindDEGene}
\alias{RunFindDEGene}
\title{Find differential expressed gene}
\usage{
RunFindDEGene(new_data, data_meta)
}
\arguments{
\item{new_data}{data input}

\item{data_meta}{meta input}
}
\value{
A numeric vector of log density.
}
\description{
Creating poisson vector.
}
\details{
Input an integer and return the log density of a poisson distribution with lambda equals the input integer
}
