% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSubclassClassify.R
\name{RunSubclassClassify}
\alias{RunSubclassClassify}
\title{Classify class and subclass}
\usage{
RunSubclassClassify(new_data, ref_data, ref_meta)
}
\arguments{
\item{new_data}{data input}

\item{ref_data}{data ref}

\item{ref_meta}{meta input}
}
\value{
A numeric vector of log density.
}
\description{
Creating poisson vector.
}
\details{
Input an integer and return the log density of a poisson distribution with lambda equals the input integer
}
