% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetCellTypeName.R
\name{GetCellTypeName}
\alias{GetCellTypeName}
\title{Get cell type name}
\usage{
GetCellTypeName(input_data, cell_type_file, all_marker_list)
}
\arguments{
\item{input_data}{data input}

\item{cell_type_file}{meta input}

\item{all_marker_list}{de gene input}
}
\value{
A numeric vector of log density.
}
\description{
Creating poisson vector.
}
\details{
Input an integer and return the log density of a poisson distribution with lambda equals the input integer
}
