% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunFindDEGene.R
\name{RunFindDEGene}
\alias{RunFindDEGene}
\title{Find differential expressed gene}
\usage{
RunFindDEGene(new_data, data_meta)
}
\arguments{
\item{new_data}{Data input, single cell expression matrix.}

\item{data_meta}{Cell type classification result from RunSubclassClassify.}
}
\value{
A list with three files contains DE gene reuslt.
}
\description{
Find differential expressed gene.
}
