% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunSubclassClassify.R
\name{RunSubclassClassify}
\alias{RunSubclassClassify}
\title{Classify class and subclass}
\usage{
RunSubclassClassify(
  new_data,
  group_meta,
  ref_data,
  ref_meta,
  method = "Seurat"
)
}
\arguments{
\item{new_data}{Data input, single cell expression data.}

\item{group_meta}{meta input}

\item{ref_data}{Reference expression data for classication.}

\item{ref_meta}{Reference metadata, must include column name with class_label and subclass_label.}

\item{method}{Method to run classification, can be choosen from Seurat or SingleR. Default is Seurat.}
}
\value{
A dataframe with annotation of cell type major class and subclass.
}
\description{
Classify class and subclass.
}
